Refactor shader types and make FFmpeg linking portable
